Changes on the path of the files packge
