+++ src/Pokemons/Pokemon.java
@@ -1,7 +1,8 @@
-package Pokemons;
+package src.Pokemons;
 
 import java.util.ArrayList;
-import Actions.Attack;
+import Pokemons.Type;
+import src.Actions.Attack;
 
 public abstract class Pokemon {
     private String name;
